requirements.txt
# ==============================
# Core Dependencies
# ==============================
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
requests==2.31.0
pandas==2.2.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiosqlite==0.19.0
requests==2.31.0

# ==============================
# Authentication & Security
# ==============================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.1

# ==============================
# Development & Testing
# ==============================
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
pylint>=2.17.0
mypy>=1.5.0

# ==============================
# Security Tools
# ==============================
bandit>=1.7.5
# safety intentionally removed due to Pydantic conflict; re-add only if necessary
# safety>=2.3.0

# ==============================
# Documentation
# ==============================
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0
